backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.10.0
pymysql>=1.0.0
cryptography>=40.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
python-multipart>=0.0.5
python-dotenv>=1.0.0
stripe>=6.0.0
paypalrestsdk>=1.13.0
redis>=4.0.0
celery>=5.0.0
boto3>=1.26.0
pillow>=9.0.0
email-validator>=2.0.0
jinja2>=3.0.0
aiofiles>=23.0.0
httpx>=0.24.0
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-cov>=4.0.0
faker>=19.0.0
slowapi>=0.1.0
python-slugify>=8.0.0
